Refactor ITs
