+++ src/it/java/org/jboss/tools/intellij/openshift/utils/odo/OdoCliTest.java
@@ -11,42 +11,35 @@
 package org.jboss.tools.intellij.openshift.utils.odo;
 
 import io.fabric8.kubernetes.client.ConfigBuilder;
-import io.fabric8.openshift.api.model.Project;
 import io.fabric8.openshift.client.DefaultOpenShiftClient;
 import io.fabric8.openshift.client.OpenShiftClient;
 import org.apache.commons.io.FileUtils;
 import org.jboss.tools.intellij.openshift.BaseTest;
 import org.junit.Before;
-import org.junit.Ignore;
-import org.junit.Test;
 
 import java.io.File;
 import java.io.IOException;
-import java.util.List;
 import java.util.Random;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertTrue;
 
-public class OdoCliTest extends BaseTest {
+public abstract class OdoCliTest extends BaseTest {
 
     public static final String COMPONENT_PATH = "src/it/projects/springboot-rest";
-    private Odo odo;
+    protected Odo odo;
 
-    private OpenShiftClient client;
+    protected OpenShiftClient client;
 
-    private Random random = new Random();
+    protected Random random = new Random();
 
-    private static final String PROJECT_PREFIX = "prj";
+    protected static final String PROJECT_PREFIX = "prj";
 
-    private static final String APPLICATION_PREFIX = "app";
+    protected static final String APPLICATION_PREFIX = "app";
 
-    private static final String COMPONENT_PREFIX = "comp";
+    protected static final String COMPONENT_PREFIX = "comp";
 
-    private static final String SERVICE_PREFIX = "srv";
+    protected static final String SERVICE_PREFIX = "srv";
 
-    private static final String STORAGE_PREFIX = "stor";
+    protected static final String STORAGE_PREFIX = "stor";
 
     @Before
     public void init() throws Exception {
@@ -59,239 +52,19 @@ public class OdoCliTest extends BaseTest {
         Thread.sleep(1000);
     }
 
-    private void createProject(String project) throws IOException, InterruptedException {
+    protected void createProject(String project) throws IOException, InterruptedException {
         odo.createProject(project);
         pause();
     }
 
-    private void createComponent(String project, String application, String component) throws IOException, InterruptedException {
+    protected void createComponent(String project, String application, String component, boolean push) throws IOException, InterruptedException {
         createProject(project);
         FileUtils.deleteDirectory(new File(COMPONENT_PATH, ".odo"));
-        odo.createComponentLocal(project, application, "java", "8", component, new File(COMPONENT_PATH).getAbsolutePath(), false);
+        odo.createComponentLocal(project, application, "java", "8", component, new File(COMPONENT_PATH).getAbsolutePath(), push);
     }
 
-    private void createStorage(String project, String application, String component, String storage) throws IOException, InterruptedException {
-        createComponent(project, application, component);
+    protected void createStorage(String project, String application, String component, boolean push, String storage) throws IOException, InterruptedException {
+        createComponent(project, application, component, push);
         odo.createStorage(project, application, COMPONENT_PATH, component, storage, "/tmp", "1Gi");
-    }
-
-    @Test
-    public void checkCreateProject() throws IOException, InterruptedException {
-        String project = PROJECT_PREFIX + random.nextInt();
-        try {
-            createProject(project);
-        } finally {
-            try {
-                odo.deleteProject(project);
-            } catch (IOException e) {}
-        }
-    }
-
-
-    @Test
-    public void checkGetComponentTypes() throws IOException {
-        List<ComponentType> components = odo.getComponentTypes();
-        assertTrue(components.size() > 0);
-    }
-
-    @Test
-    public void checkGetServiceTemplates() throws IOException {
-        List<ServiceTemplate> services = odo.getServiceTemplates();
-        assertTrue(services.size() > 0);
-    }
-
-    @Test
-    public void checkListProjects() throws IOException, InterruptedException {
-        String project = PROJECT_PREFIX + random.nextInt();
-        try {
-            createProject(project);
-            List<Project> projects = odo.getProjects(client);
-            assertTrue(projects.size() > 0);
-        } finally {
-            try {
-                odo.deleteProject(project);
-            } catch (IOException e) {}
-        }
-    }
-
-    @Test
-    public void checkCreateComponent() throws IOException, InterruptedException {
-        String project = PROJECT_PREFIX + random.nextInt();
-        String application = APPLICATION_PREFIX + random.nextInt();
-        String component = COMPONENT_PREFIX + random.nextInt();
-        try {
-            createComponent(project, application, component);
-        } finally {
-            try {
-                odo.deleteProject(project);
-            } catch (IOException e) {}
-        }
-    }
-
-    @Test
-    public void checkCreateAndDeleteComponent() throws IOException, InterruptedException {
-        String project = PROJECT_PREFIX + random.nextInt();
-        String application = APPLICATION_PREFIX + random.nextInt();
-        String component = COMPONENT_PREFIX + random.nextInt();
-        try {
-            createComponent(project, application, component);
-            odo.deleteComponent(project, application, COMPONENT_PATH, component, false);
-        } finally {
-            try {
-                odo.deleteProject(project);
-            } catch (IOException e) {}
-        }
-    }
-
-    @Test
-    public void checkCreateComponentAndCreateURL() throws IOException, InterruptedException {
-        String project = PROJECT_PREFIX + random.nextInt();
-        String application = APPLICATION_PREFIX + random.nextInt();
-        String component = COMPONENT_PREFIX + random.nextInt();
-        try {
-            createComponent(project, application, component);
-            odo.createURL(project, application, COMPONENT_PATH, component, "url1", 8080);
-            List<URL> urls = odo.listURLs(project, application, COMPONENT_PATH, component);
-            assertEquals(1, urls.size());
-        } finally {
-            try {
-                odo.deleteProject(project);
-            } catch (IOException e) {}
-        }
-    }
-
-    @Test
-    public void checkCreateComponentAndCreateAndDeleteURL() throws IOException, InterruptedException {
-        String project = PROJECT_PREFIX + random.nextInt();
-        String application = APPLICATION_PREFIX + random.nextInt();
-        String component = COMPONENT_PREFIX + random.nextInt();
-        try {
-            createComponent(project, application, component);
-            odo.createURL(project, application, COMPONENT_PATH, component, null, 8080);
-            List<URL> urls = odo.listURLs(project, application, COMPONENT_PATH, component);
-            assertEquals(1, urls.size());
-            odo.deleteURL(project, application, COMPONENT_PATH, component, urls.get(0).getName());
-            urls = odo.listURLs(project, application, COMPONENT_PATH, component);
-            assertEquals(0, urls.size());
-        } finally {
-            try {
-                odo.deleteProject(project);
-            } catch (IOException e) {}
-        }
-    }
-
-    @Test
-    @Ignore("not yet supported by odo")
-    public void checkCreateComponentAndLinkService() throws IOException, InterruptedException {
-        String project = PROJECT_PREFIX + random.nextInt();
-        String application = APPLICATION_PREFIX + random.nextInt();
-        String component = COMPONENT_PREFIX + random.nextInt();
-        String service = SERVICE_PREFIX + random.nextInt();
-        try {
-            createComponent(project, application, component);
-            odo.createService(project, application, "postgresql-persistent", "default", service);
-            odo.link(project, application, component, COMPONENT_PATH, service, null);
-        } finally {
-            try {
-                odo.deleteProject(project);
-            } catch (IOException e) {}
-        }
-    }
-
-    @Test
-    public void checkCreateComponentAndCreateStorage() throws IOException, InterruptedException {
-        String project = PROJECT_PREFIX + random.nextInt();
-        String application = APPLICATION_PREFIX + random.nextInt();
-        String component = COMPONENT_PREFIX + random.nextInt();
-        String storage = STORAGE_PREFIX + random.nextInt();
-        try {
-            createStorage(project, application, component, storage);
-        } finally {
-            try {
-                odo.deleteProject(project);
-            } catch (IOException e) {}
-        }
-    }
-
-    @Test
-    public void checkCreateComponentAndCreateDeleteStorage() throws IOException, InterruptedException {
-        String project = PROJECT_PREFIX + random.nextInt();
-        String application = APPLICATION_PREFIX + random.nextInt();
-        String component = COMPONENT_PREFIX + random.nextInt();
-        String storage = STORAGE_PREFIX + random.nextInt();
-        try {
-            createStorage(project, application, component, storage);
-            odo.deleteStorage(project, application, COMPONENT_PATH, component, storage);
-        } finally {
-            try {
-                odo.deleteProject(project);
-            } catch (IOException e) {}
-        }
-    }
-
-    @Test
-    public void checkCreateComponentAndListURLs() throws IOException, InterruptedException {
-        String project = PROJECT_PREFIX + random.nextInt();
-        String application = APPLICATION_PREFIX + random.nextInt();
-        String component = COMPONENT_PREFIX + random.nextInt();
-        try {
-            createComponent(project, application, component);
-            List<URL> urls = odo.listURLs(project, application, COMPONENT_PATH, component);
-            assertEquals(0, urls.size());
-        } finally {
-            try {
-                odo.deleteProject(project);
-            } catch (IOException e) {}
-        }
-    }
-
-    @Test
-    public void checkCreateService() throws IOException, InterruptedException {
-        String project = PROJECT_PREFIX + random.nextInt();
-        String application = APPLICATION_PREFIX + random.nextInt();
-        String service = SERVICE_PREFIX + random.nextInt();
-        try {
-            createProject(project);
-            odo.createService(project, application, "postgresql-persistent", "default", service);
-        } finally {
-            try {
-                odo.deleteProject(project);
-            } catch (IOException e) {}
-        }
-    }
-
-    @Test
-    public void checkCreateServiceAndGetTemplate() throws IOException, InterruptedException {
-        String project = PROJECT_PREFIX + random.nextInt();
-        String application = APPLICATION_PREFIX + random.nextInt();
-        String service = SERVICE_PREFIX + random.nextInt();
-        try {
-            createProject(project);
-            odo.createService(project, application, "postgresql-persistent", "default", service);
-            String template = odo.getServiceTemplate(client, project, application, service);
-            assertNotNull(template);
-            assertEquals("postgresql-persistent", template);
-        } finally {
-            try {
-                odo.deleteProject(project);
-            } catch (IOException e) {}
-        }
-    }
-
-    @Test
-    @Ignore("not yet supported by odo")
-    public void checkCreateDeleteService() throws IOException, InterruptedException {
-        String project = PROJECT_PREFIX + random.nextInt();
-        String application = APPLICATION_PREFIX + random.nextInt();
-        String service = SERVICE_PREFIX + random.nextInt();
-        try {
-            createProject(project);
-            odo.createService(project, application, "postgresql-persistent", "default", service);
-            odo.deleteService(project, application, service);
-        } finally {
-            try {
-                odo.deleteProject(project);
-            } catch (IOException e) {}
-        }
     }
 }
